Move ExperimentData to framework
