+++ qiskit_experiments/framework/base_analysis.py
@@ -22,7 +22,7 @@ from qiskit.providers.options import Options
 from qiskit_experiments.exceptions import AnalysisError
 from qiskit_experiments.database_service import DbAnalysisResultV1
 from qiskit_experiments.database_service.device_component import Qubit
-from qiskit_experiments.experiment_data import ExperimentData
+from qiskit_experiments.framework.experiment_data import ExperimentData
 
 
 class BaseAnalysis(ABC):
