+++ qiskit_experiments/library/tomography/tomography_analysis.py
@@ -25,8 +25,7 @@ from qiskit.quantum_info.operators.base_operator import BaseOperator
 from qiskit.quantum_info.operators.channel.quantum_channel import QuantumChannel
 
 from qiskit_experiments.exceptions import AnalysisError
-from qiskit_experiments.framework import BaseAnalysis, Options
-from qiskit_experiments.experiment_data import AnalysisResultData
+from qiskit_experiments.framework import BaseAnalysis, Options, AnalysisResultData
 from .fitters import (
     linear_inversion,
     scipy_linear_lstsq,
